chore: Rename `staff_qr_validator.dart`
